MCQService/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.12.4
transformers==4.36.0
torch==2.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3
boto3==1.34.0
python-multipart==0.0.6
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
sentencepiece==0.1.99
